% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.r
\name{clump_gwasvcf}
\alias{clump_gwasvcf}
\title{Perform LD clumping}
\usage{
clump_gwasvcf(
  vcf,
  clump_kb = 1000,
  clump_r2 = 0.001,
  clump_p = 5e-08,
  pop = NULL,
  bfile = NULL,
  plink_bin = NULL,
  access_token = NULL
)
}
\arguments{
\item{vcf}{VCF file or VCF object}

\item{clump_kb}{Clumping kb window. Default is very strict, \code{10000}}

\item{clump_r2}{Clumping r2 threshold. Default is very strict, \code{0.001}}

\item{clump_p}{Clumping sig level for index variants. Default = \code{1}
(i.e. no threshold)}

\item{pop}{Super-population to use as reference panel. Default = \code{"EUR"}.
Options are \code{"EUR"}, \code{"SAS"}, \code{"EAS"}, \code{"AFR"}, \code{"AMR"}.
\code{'legacy'} also available - which is a previously used verison of the EUR
panel with a slightly different set of markers}

\item{bfile}{If this is provided then will use the API. Default = \code{NULL}}

\item{plink_bin}{If \code{NULL} and bfile is not \code{NULL} then will detect packaged
plink binary for specific OS. Otherwise specify path to plink binary.
Default = \code{NULL}}

\item{access_token}{Google OAuth2 access token. Used to authenticate level
of access to data}
}
\value{
data frame of clumped results
}
\description{
Perform LD clumping
}
