% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cojo.r
\name{cojo_sumstat_file}
\alias{cojo_sumstat_file}
\title{Write vcf file to cojo sumstat file}
\usage{
cojo_sumstat_file(vcffile, outfile)
}
\arguments{
\item{vcffile}{Path to vcf file}

\item{outfile}{Path to output file}
}
\value{
vcf object
}
\description{
Write vcf file to cojo sumstat file
}
