% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gassocplot.r
\name{coloc_to_gassocplot}
\alias{coloc_to_gassocplot}
\title{Convert coloc dataset to gassocplot dataset}
\usage{
coloc_to_gassocplot(coloclist, bfile = NULL, plink_bin = NULL)
}
\arguments{
\item{coloclist}{Output from *_to_coloc}

\item{bfile}{If number of SNPs > 500 then need to provide your own LD reference panel. Provide plink dataset here.}

\item{plink_bin}{If number of SNPs > 500 then need to provide your own LD reference panel. Provide plink executable here}
}
\value{
List to feed into gassocplot
}
\description{
Convert coloc dataset to gassocplot dataset
}
