% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.r
\name{gwasvcf_to_coloc4}
\alias{gwasvcf_to_coloc4}
\title{Generate coloc v4  dataset from vcf files}
\usage{
gwasvcf_to_coloc4(
  vcf1,
  vcf2,
  chrompos,
  bfile,
  plink_bin = getOption("tools_plink")
)
}
\arguments{
\item{vcf1}{VCF object or path to vcf file}

\item{vcf2}{VCF object or path to vcf file}

\item{chrompos}{Character of chr:pos1-pos2}
}
\value{
List of datasets to feed into coloc (v4)
}
\description{
Generate coloc v4  dataset from vcf files
}
