% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapr.r
\name{gwasvcf_to_finemapr}
\alias{gwasvcf_to_finemapr}
\title{Generate data for analysis in various finemapping methods}
\usage{
gwasvcf_to_finemapr(
  region,
  vcf,
  bfile = NULL,
  plink_bin = getOption("tools_plink")
)
}
\arguments{
\item{region}{Region of the genome to extract eg 1:109317192-110317192"}

\item{vcf}{list of paths to gwasvcf files}

\item{bfile}{If this is provided then will use the API. Default = NULL}

\item{plink_bin}{If null and bfile is not null then will detect packaged plink binary for specific OS. Otherwise specify path to plink binary. Default = NULL}
}
\value{
Each vcf will be a list of z score data, ld matrix, and sample size
}
\description{
Uses the finemapr package https://github.com/variani/finemapr
}
