% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwcoco.r
\name{gwasvcf_to_pwcoco}
\alias{gwasvcf_to_pwcoco}
\title{Write files for PWCoCo where data are read from two VCF objects or files.}
\usage{
gwasvcf_to_pwcoco(vcf1, vcf2, chrompos, type1 = NULL, type2 = NULL, outfile)
}
\arguments{
\item{vcf1}{VCF object or path to VCF file}

\item{vcf2}{VCF object or path to VCF file}

\item{chrompos}{Character of the format chr:pos1-pos2}

\item{type1}{How to treat vcffile1 for coloc, either \code{"quant"} or \code{"cc"}}

\item{type2}{How to treat vcffile2 for coloc, either \code{"quant"} or \code{"cc"}}

\item{outfile}{Path to output files, without file ending}
}
\value{
\code{0} if success, \code{1} if there was a problem
}
\description{
Write files for PWCoCo where data are read from two VCF objects or files.
}
