% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR.r
\name{gwasvcf_to_TwoSampleMR}
\alias{gwasvcf_to_TwoSampleMR}
\title{Create exposure or outcome data format for TwoSampleMR from vcf}
\usage{
gwasvcf_to_TwoSampleMR(vcf, type = "exposure")
}
\arguments{
\item{vcf}{VCF object}

\item{type}{=\code{"exposure"} or \code{"outcome"}}
}
\value{
data frame
}
\description{
Create exposure or outcome data format for TwoSampleMR from vcf
}
