% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{harmonise}
\alias{harmonise}
\title{Generic harmonisation function}
\usage{
harmonise(
  chr1,
  pos1,
  ref1,
  alt1,
  chr2,
  pos2,
  ref2,
  alt2,
  rsid2 = NULL,
  indel_recode = FALSE,
  strand_flip = FALSE
)
}
\arguments{
\item{chr1}{Vector}

\item{pos1}{Vector}

\item{ref1}{Vector}

\item{alt1}{Vector}

\item{chr2}{Vector}

\item{pos2}{Vector}

\item{ref2}{Vector}

\item{alt2}{Vector}

\item{rsid2}{Optional vector}

\item{indel_recode}{=FALSE. If TRUE then attempts to recode D/I}

\item{strand_flip}{=FALSE. If TRUE then attempts to flip strand when alignment is not otherwise possible}
}
\value{
Dataframe of outcomes
}
\description{
Assumes ref and alt alleles available for target and reference datasets, and uses chr:pos for alignment
}
\details{
0: stick
1: swap
2: rename indel
3: rename indel and swap
4: flip
5: flip and swap
6: drop (no match)
7: drop (no reference)
}
