% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{harmonise_against_ref}
\alias{harmonise_against_ref}
\title{Harmonise gwas alleles to be same as reference}
\usage{
harmonise_against_ref(gwas, reference)
}
\arguments{
\item{gwas}{data.frame indicating GWAS}

\item{reference}{data.frame indicating reference dataset}
}
\value{
data frame with attributes
}
\description{
Harmonise gwas alleles to be same as reference
}
