% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{ieu_to_TwoSampleMR}
\alias{ieu_to_TwoSampleMR}
\title{Convert output from query to TwoSampleMR format}
\usage{
ieu_to_TwoSampleMR(x, type = "exposure")
}
\arguments{
\item{x}{Output from ieugwasr query e.g. associations, tophits, phewas}

\item{type}{"exposure" (default) or "outcome"}
}
\value{
data frame
}
\description{
Convert output from query to TwoSampleMR format
}
