% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwcoco.r
\name{ieugwasr_to_pwcoco}
\alias{ieugwasr_to_pwcoco}
\title{Write files for PWCoCo where data are read from the OpenGWAS DB.}
\usage{
ieugwasr_to_pwcoco(id1, id2, chrompos, type1 = NULL, type2 = NULL, outfile)
}
\arguments{
\item{id1}{ID for trait 1}

\item{id2}{ID for trait 2}

\item{chrompos}{Character of the format chr:pos1-pos2}

\item{type1}{How to treat vcffile1 for coloc, either "quant" or "cc"}

\item{type2}{How to treat vcffile2 for coloc, either "quant" or "cc"}

\item{outfile}{Path to output files, without file ending

return 0 if success, 1 if there was a problem}
}
\description{
Write files for PWCoCo where data are read from the OpenGWAS DB.
}
