% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR.r
\name{ieugwasr_to_TwoSampleMR}
\alias{ieugwasr_to_TwoSampleMR}
\title{Convert output from query to TwoSampleMR format}
\usage{
ieugwasr_to_TwoSampleMR(x, type = "exposure")
}
\arguments{
\item{x}{Output from ieugwasr query e.g. associations, tophits, phewas}

\item{type}{\code{"exposure"} (default) or \code{"outcome"}}
}
\value{
data frame
}
\description{
Convert output from query to TwoSampleMR format
}
