% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cojo.r
\name{map_variants_to_regions}
\alias{map_variants_to_regions}
\title{For a set of variants map to LD regions}
\usage{
map_variants_to_regions(chrpos, pop)
}
\arguments{
\item{chrpos}{Array of chr:pos}

\item{pop}{EUR, AFR or ASN}
}
\description{
LD regions defined here https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4731402/
}
