% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwcoco.r
\name{pwcoco}
\alias{pwcoco}
\title{Perform pair-wise conditional and colocalisation analysis using PWCoCo}
\usage{
pwcoco(
  id1,
  id2,
  bfile,
  chrompos,
  pwcoco,
  type1 = NULL,
  type2 = NULL,
  workdir = tempdir()
)
}
\arguments{
\item{id1}{Path to vcffile or ID for trait1}

\item{id2}{Path to vcffile2 or ID for trait2}

\item{bfile}{LD reference panel in Plink format (.bed, .bim, .fam)}

\item{chrompos}{Chromosome position (format: chr:pos1-pos2) region of
interest}

\item{pwcoco}{Path to pwcoco binary}

\item{type1}{How to treat vcffile1 for coloc, either \code{"quant"} or \code{"cc"}}

\item{type2}{How to treat vcffile2 for coloc, either \code{"quant"} or \code{"cc"}}

\item{workdir}{Location to store files, default=\code{tempdir()}}
}
\value{
List of colocalisation results
}
\description{
For a list of fine-mapped rsids, will assign to regions and generate
colocalisation results for conditionally independent summary stats for each
rsid
}
