% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{read_gwas}
\alias{read_gwas}
\title{Read in GWAS dataset}
\usage{
read_gwas(filename, skip, delimiter, gzipped, snp, nea, ea, ea_af, effect,
  se, pval, n, info, z)
}
\arguments{
\item{filename}{<what param does>}

\item{skip}{<what param does>}

\item{delimiter}{<what param does>}

\item{gzipped}{<what param does>}

\item{snp}{<what param does>}

\item{nea}{<what param does>}

\item{ea}{<what param does>}

\item{ea_af}{<what param does>}

\item{effect}{<what param does>}

\item{se}{<what param does>}

\item{pval}{<what param does>}

\item{n}{<what param does>}

\item{info}{<what param does>}

\item{z}{<what param does>}
}
\value{
data frame with log attributes
}
\description{
Read in GWAS dataset
}
