% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{read_reference}
\alias{read_reference}
\title{Read in reference dataset}
\usage{
read_reference(
  reference_file,
  rsid = NULL,
  chrompos = NULL,
  remove_dup_rsids = TRUE
)
}
\arguments{
\item{reference_file}{Reference vcf}

\item{rsid}{List of variants to read}

\item{chrompos}{List of chrompos to read}

\item{remove_dup_rsids}{=\code{TRUE} Remove duplicates from output}
}
\value{
data frame
}
\description{
Read in reference dataset
}
