% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleMR.r
\name{set_bc4_files}
\alias{set_bc4_files}
\title{Determine locations of useful reference datasets on bluecrystal4}
\usage{
set_bc4_files()
}
\description{
This is a convenience function for members at the University of Bristol
to automatically set file locations for various reference datasets. It relates
only to paths on bc4
}
