% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susieR.r
\name{susieR_pipeline}
\alias{susieR_pipeline}
\title{Perform fine mapping pipeline using susieR}
\usage{
susieR_pipeline(
  vcffile,
  bfile,
  plink_bin,
  pop,
  threads = 1,
  clump_kb = 1000,
  clump_r2 = 0.001,
  clump_p = 5e-08,
  ...
)
}
\arguments{
\item{vcffile}{Path to vcf file}

\item{bfile}{Path to ld reference panel}

\item{plink_bin}{Path to plink}

\item{pop}{EUR, ASN or AFR}

\item{threads}{Numeric indicating number of threads. Default \code{1}}

\item{clump_kb}{Numeric indicating kilobases used in clumping.
Default \code{1000}}

\item{clump_r2}{Numeric indicating R-squared used in clumping.
Default \code{0.001}}

\item{clump_p}{Numeric indicating clumping p-value. Default \code{5e-8}}

\item{...}{Optional arguments to be passed to \code{\link[susieR:susie_rss]{susieR::susie_rss}}}
}
\value{
List
}
\description{
Clumps data, then maps those to LD representative regions. Within each
detected LD representative region, fine mapping is performed.
}
