% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.r
\name{gwasvcf_to_coloc}
\alias{gwasvcf_to_coloc}
\title{Generate coloc dataset from vcf files}
\usage{
gwasvcf_to_coloc(vcf1, vcf2, chrompos)
}
\arguments{
\item{vcf1}{VCF object or path to vcf file}

\item{vcf2}{VCF object or path to vcf file}

\item{chrompos}{Character of chr:pos1-pos2}
}
\value{
List of datasets to feed into coloc
}
\description{
Generate coloc dataset from vcf files
}
