% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapr.r
\name{gwasvcf_to_finemapr}
\alias{gwasvcf_to_finemapr}
\title{Generate data for fine mapping analysis}
\usage{
gwasvcf_to_finemapr(
  region,
  vcf,
  bfile,
  plink_bin = genetics.binaRies::get_plink_binary(),
  threads = 1
)
}
\arguments{
\item{region}{Region of the genome to extract eg 1:109317192-110317192". Can be array}

\item{vcf}{Path to VCF file or VCF object}

\item{bfile}{LD reference panel}

\item{plink_bin}{Path to plink. Default = genetics.binaRies::get_plink_binary()}

\item{threads}{Number of threads to run in parallel. Default=1}
}
\value{
List of datasets for finemapping
}
\description{
For a given region and VCF file, extracts the variants in the region along with LD matrix from a reference panel
}
