% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{read_gwas}
\alias{read_gwas}
\title{Read in GWAS dataset}
\usage{
read_gwas(
  filename,
  skip,
  delimiter,
  gzipped,
  snp,
  nea,
  ea,
  ea_af,
  effect,
  se,
  pval,
  n,
  info,
  z
)
}
\arguments{
\item{filename}{Filename}

\item{skip}{Option passed to \code{\link[data.table:fread]{data.table::fread}}}

\item{delimiter}{The separator between columns in the file}

\item{gzipped}{Logical indicating whether the file is compressed in gz format}

\item{snp}{Numeric indicating column number with SNP name}

\item{nea}{Numeric indicating column number with NEA}

\item{ea}{Numeric indicating column number with effect allele}

\item{ea_af}{Numeric indicating column number with effect allele
allele frequency}

\item{effect}{Numeric indicating column number with effect}

\item{se}{Numeric indicating column number with standard error of effect}

\item{pval}{Numeric indicating column number with p-value of effect}

\item{n}{Numeric indicating column number with sample size}

\item{info}{Numeric indicating column number with info}

\item{z}{Numeric indicating column number with z statistic of effect}
}
\value{
data frame with log attributes
}
\description{
Read in GWAS dataset
}
