% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susieR.r
\name{susieR_pipeline}
\alias{susieR_pipeline}
\title{Perform fine mapping pipeline using susieR}
\usage{
susieR_pipeline(
  vcffile,
  bfile,
  plink_bin,
  pop,
  threads = 1,
  clump_kb = 1000,
  clump_r2 = 0.001,
  clump_p = 5e-08,
  ...
)
}
\arguments{
\item{vcffile}{Path to vcf file}

\item{bfile}{Path to ld reference panel}

\item{plink_bin}{Path to plink}

\item{pop}{EUR, ASN or AFR}

\item{clump_kb}{<what param does>}

\item{clump_r2}{<what param does>}

\item{clump_p}{<what param does>}

\item{...}{Optional arguments to be passed to susie_rss}
}
\value{
List
}
\description{
Clumps data, then maps those to LD representative regions. Within each detected LD representative region, fine mapping is performed
}
