% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.r
\name{write_out}
\alias{write_out}
\title{Create format for HPC pipeline}
\usage{
write_out(harmonised, path)
}
\arguments{
\item{harmonised}{Output from \code{\link{harmonise_against_ref}}}

\item{path}{Path to write out json file and txt file}
}
\description{
Takes raw files and aligns them to reference. Important if files don't have
chr:pos already
}
